requirements.txt
requests>=2.31.0
feedparser>=6.0.11
Flask>=2.2
python-dotenv>=1.0
newspaper3k>=0.2.8
spacy>=3.0.0
transformers>=4.30.0
torch>=1.13.0
redis>=4.5.0
